Correct summary stats in GUI
Solves problem that if not all numerical variables are changed, the
summary view in the GUI was wrong (table was filled up and duplicated
results). Proposed solution to adapt the print function: dt is NULL if
no changes, if any variable changes it also includes the summary
statistics for the remaining variables (without changes). Alternative
would be to only show the changed variable sin the GUI