Eigenvalue computation can fail setup SDC problem
Upon setting up an sdc problem with numerical keys, the eigenvalues are
automatically computed. If these fail, the sdc object isn’t created.
Therefore, I suggest to set the difference in eigenvalues to NA if the
computation fails. The function eigen() fails, which is probably
because it’s impossibkl to compute the eigenvalues (e.g., very many
missing values in some variables)