add parameter `shiny.server` to `sdcApp`
In case this parameter is set to TRUE (not default), sdcApp will return the app as an
object instead of incoking it. As the name suggests, this is useful for deploying
the app via shiny-server.

These changes also lead to the `.GlobalEnv` not being used by the app (even if the `shiny.server`
flag is set to `FALSE`). Instead, local variables are saved in the app object.

Parameters are now passed via `shinyOptions/getShinyOption`. Again, this keeps the global
environment clean compared to the previous approach, which used the global environment.